refactor(player): move MydiaLogo out of the app shell
