+++ player/lib/presentation/screens/home_screen.dart
@@ -18,6 +18,7 @@ import '../widgets/shimmer_card.dart';
 import '../../core/layout/breakpoints.dart';
 import '../../core/theme/colors.dart';
 import '../widgets/app_shell.dart';
+import '../widgets/mydia_logo.dart';
 import 'home/home_controller.dart';
 
 String _resumeSuffix({
